adapt page reservas to next
