+++ src/components/button/style.module.scss
@@ -39,3 +39,18 @@
 		color: #C96000;
 	}
 }
+
+.disableButton {
+	justify-content: center;
+	align-items: center;
+	display: flex;
+	padding: 12px 24px;
+	font-size: 16px;
+	font-family: inherit;
+	background-color: #343434;
+	cursor: default;
+	color: #FFF;
+	letter-spacing: 1.2;
+	border: 1px solid transparent;
+	transition: all 0.3s;
+}
