+++ src/components/reservas/style.module.scss
@@ -1,0 +1,170 @@
+.modalOverlay {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	position: fixed;
+	top: 0;
+	left: 0;
+	right: 0;
+	bottom: 0;
+	background: rgba(0, 0, 0, 0.5);
+}
+
+.modal {
+	color:#000000;
+	background: #fff;
+	border-radius: 10px;
+	max-width: 400px;
+	padding: 20px;
+	text-align: center;
+	position: relative;
+}
+
+.modalContent {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+
+	label {
+		display: block;
+		margin-bottom: 10px;
+	}
+
+	input {
+		width: 100%;
+		padding: 5px;
+		margin-bottom: 10px;
+		border: 1px solid #ccc;
+		border-radius: 5px;
+	}
+
+	p {
+		word-break: break-all;
+	}
+}
+
+.closeModalButton {
+	width: 5px;
+	height: 5px;
+	position: absolute;
+	top: 10px;
+	right: 10px;
+	font-size: 20px;
+	cursor: pointer;
+}
+
+.findTablesButton {
+	background-color: #ff6600;
+	color: #fff;
+	border-color: #ff6600;
+	padding: 10px 20px;
+	border-radius: 5px;
+	cursor: pointer;
+	position: center;
+	bottom: 10px;
+	left: 50%;
+}
+
+.findTablesButton:hover {
+	color: #fff;
+	background-color: #c95000;
+	border-color: #c95000;
+}
+
+/* Estilos para a seleção de areas */
+.areasContainer {
+	display: flex;
+	flex-direction: row;
+	gap: 20px;
+	margin-bottom: 20px;
+
+	button {
+		cursor: pointer;
+		border: 0;
+		font-size: 16px;
+		border-color: #fff;
+		background-color: #fff;
+
+		&:hover {
+			color: #c95000;
+			transition: all 0.3s;
+		}
+
+		&:focus {
+			color: #c95000;
+		}
+	}
+}
+
+.area {
+	background: #ff6600;
+	color: #fff;
+	border: 2px solid #c95000;
+	padding: 10px;
+	border-radius: 10px;
+	cursor: pointer;
+	font-weight: bold;
+	font-size: 16px;
+}
+
+.customLayout {
+	display: flex;
+	justify-content: space-between;
+}
+
+.leftColumn {
+	display: flex;
+	flex-direction: column;
+	margin: 50px 50px 0 50px;
+	text-align: center;
+	align-items: center;
+}
+
+.rightColumn {
+	flex: 1;
+	margin: 0 50px 0 50px;
+	text-align: right;
+}
+
+.title{
+	text-align: center;
+	padding-top: 50px;
+}
+
+@media screen and (max-width: 1320px) {
+	.customLayout {
+		display: flex;
+		flex-direction: column;
+		align-items: center;
+	}
+
+	.image {
+		width: 860px;
+		height: 320px;
+		object-fit: cover;
+	}
+}
+
+@media screen and (max-width: 900px) {
+	.image {
+		width: 750px;
+		height: 320px;
+		object-fit: cover;
+	}
+}
+
+@media screen and (max-width: 780px) {
+	.image {
+		width: 550px;
+		height: 320px;
+		object-fit: cover;
+	}
+}
+
+@media screen and (max-width: 580px) {
+	.image {
+		width: 450px;
+		height: 320px;
+		object-fit: cover;
+	}
+}
