--- src/pages/reservas.js
@@ -1,251 +1,0 @@
-import React, { useState } from 'react';
-import Header from '@/components/header/header';
-import Footer from '@/components/footer/footer';
-
-const Reservas = () => {
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [page, setPage] = useState('form'); // 'form' para o formulário, 'areas' para a seleção de areas, 'confirmacao' para a confirmação
-  const areas = [
-    { name: 'Interna', color: '#c95000' }, // Cor azul para Interna
-    { name: 'Externa', color: '#c95000' }, // Cor verde para Externa
-    { name: 'Indiferente', color: '#c95000' }, // Cor laranja para Indiferente
-  ];
-  const [areasSelecionadas, setAreasSelecionadas] = useState([]);
-  const [formData, setFormData] = useState({
-    quantidadePessoas: '',
-    data: '',
-    horario: '',
-    email: '',
-  });
-
-  const validateEmail = (email) => {
-    const regex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-    return regex.test(email);
-  };
-
-  const validateForm = () => {
-    // Validação da quantidade de pessoas
-    if (formData.quantidadePessoas < 1 || formData.quantidadePessoas > 10) {
-      alert('A quantidade de pessoas deve ser entre 1 e 10.');
-      return false; // Impede a transição
-    }
-  
-    // Validação da data (verifica se é uma data válida)
-    const selectedDate = new Date(formData.data);
-    if (isNaN(selectedDate.getTime()) || selectedDate < new Date()) {
-      alert('Selecione uma data válida no futuro.');
-      return false; // Impede a transição
-    }
-  
-    // Validação do horário (verifica se está dentro do horário desejado)
-    if (formData.horario < '18:00' || formData.horario > '22:30') {
-      alert('Selecione um horário entre 18:00 e 22:30.');
-      return false; // Impede a transição
-    }
-  
-    // Validação do e-mail (verifica se é um e-mail válido)
-    if (!validateEmail(formData.email)) {
-      alert('Insira um e-mail válido.');
-      return false; // Impede a transição
-    }
-  
-    return true; // Permite a transição
-  };
-
-  const openModal = () => {
-    setIsModalOpen(true);
-    setPage('form');
-  };
-
-  const closeModal = () => {
-    setIsModalOpen(false);
-  };
-  
-  const selectAreas = () => {
-    if (validateForm()) {
-      setPage('areas'); // Permite a transição apenas se a validação passar
-    }
-  };
-
-  const handleAreaSelection = (area) => {
-    if (areasSelecionadas.includes(area)) {
-      // Se a área já estiver selecionada, desmarque-a
-      setAreasSelecionadas([]);
-    } else {
-      // Se a área ainda não estiver selecionada, selecione-a e desmarque as outras áreas
-      setAreasSelecionadas([area]);
-    }
-  
-    // Adicione uma mensagem de aviso
-    //const message = areasSelecionadas.length > 0
-    //  ? `Você desmarcou a área "${areasSelecionadas[0]}" e selecionou a área "${area}"`
-    //  : `Você selecionou a área "${area}"`;
-  
-    //alert(message);
-  };
-  
-
-  const confirmarAreas = () => {
-    setPage('confirmacao');
-  };
-
-  const confirmarReserva = () => {
-    // Lógica para confirmar a reserva
-    // Você pode enviar os dados do formulário e as areas selecionadas para o servidor ou realizar ações específicas aqui
-    // Após a confirmação, você pode retornar à página de formulário ou fechar o modal, dependendo dos requisitos do seu aplicativo
-    // Por simplicidade, mostraremos um alerta de confirmação neste exemplo
-    alert('Reserva confirmada com sucesso!');
-    setIsModalOpen(false);
-    setFormData({
-      quantidadePessoas: '',
-      data: '',
-      horario: '',
-    });
-  };
-
-  const renderContent = () => {
-    if (page === 'form') {
-      return (
-        <div className="modal-content">
-          <h2>Formulário</h2>
-          <form>
-            <label>
-              Quantidade de Pessoas:*
-              <input
-                type="number"
-                required
-                min="1"
-                max="10"
-                value={formData.quantidadePessoas}
-                onChange={(e) =>
-                  setFormData({ ...formData, quantidadePessoas: e.target.value })
-                }
-              />
-            </label>
-            <label>
-              Data:*
-              <input
-                type="date"
-                required
-                min={new Date().toISOString().split('T')[0]} // Define o mínimo como a data atual
-                value={formData.data}
-                onChange={(e) => setFormData({ ...formData, data: e.target.value })}
-              />
-            </label>
-            <label>
-              Horário:*
-              <select
-                required
-                value={formData.horario}
-                onChange={(e) => setFormData({ ...formData, horario: e.target.value })}
-              >
-                <option value="18:00">18:00</option>
-                <option value="18:30">18:30</option>
-                <option value="19:00">19:00</option>
-                <option value="19:30">19:30</option>
-                <option value="20:00">20:00</option>
-                <option value="20:30">20:30</option>
-                <option value="21:00">21:00</option>
-                <option value="21:30">21:30</option>
-                <option value="22:00">22:00</option>
-                <option value="22:30">22:30</option>
-              </select>
-            </label>
-            <label>
-              E-mail de Contato:*
-              <input
-                type="e-mail"
-                required
-                value={formData.email}
-                onChange={(e) => setFormData({ ...formData, email: e.target.value })}
-              />
-            </label>
-            </form>
-            <button
-              className="find-tables-button"
-              type="button"
-              onClick={() => selectAreas()}
-            >
-              Encontrar Áreas
-            </button>
-        </div>
-      );
-    } else if (page === 'areas') {
-      return (
-        <div className="modal-content">
-          <h2>Qual a área desejada</h2>
-          <div id="areas" className="areas-container">
-            {areas.map((area) => (
-             <div
-             key={area.name}
-             className={`area ${areasSelecionadas.includes(area.name) ? 'selected' : ''}`}
-             style={{ backgroundColor: areasSelecionadas.includes(area.name) ? area.color : '' }}
-             data-area={area.name}
-             onClick={() => handleAreaSelection(area.name)}
-           >
-             {area.name}
-              </div>
-            ))}
-          </div>
-          <br/>
-          <button
-            className={`confirmar-area-button ${areasSelecionadas ? 'enabled' : ''}`}
-            onClick={confirmarAreas}
-            disabled={!areasSelecionadas}
-          >
-            Confirmar Área
-          </button>
-        </div>
-      );
-    } else if (page === 'confirmacao') {
-      return (
-        <div className="modal-content">
-          <h2>Confirmação de Reserva</h2>
-          <p>Quantidade de Pessoas: {formData.quantidadePessoas}</p>
-          <p>Data: {formData.data.split('-').reverse().join('/')}</p>
-          <p>Horário: {formData.horario}</p>
-          <p>Área Selecionada: {areasSelecionadas}</p>
-          <p>E-mail de Contato: {formData.email}</p>
-          <button className="confirmar-reserva-button" onClick={confirmarReserva}>
-            Confirmar Reserva
-          </button>
-        </div>
-      );
-    }
-  };
-
-  return (
-    <div>
-      <h1 className="title" >RESERVE JÁ!</h1>
-      <div className="custom-layout">
-        <div className="left-column">
-          <p>
-          Bem-vindo(a) ao Sushi Tech, onde a tradição japonesa encontra a inovação culinária e tecnológica! Estamos entusiasmados por você considerar a reserva de um lugar em nosso renomado restaurante japonês. 
-          <br/><br/>
-          Para garantir que sua visita ao Sushi Tech seja perfeita, convidamos você à reservar sua area com antecedência. Nossa equipe dedicada está pronta para tornar sua experiência inesquecível! 
-          <br/><br/>
-          Por favor, preencha o formulário abaixo para fazer sua reserva agora mesmo.
-          <br/><br/>
-          Mal podemos esperar para recebê-lo(a) em nossa casa de sabores japoneses de última geração. 
-          </p><br/>
-          <button className="open-modal-button" onClick={openModal}>Fazer uma reserva</button>
-          {isModalOpen && (
-          <div className="modal-overlay">
-            <div className="modal">
-              <button className="close-modal-button" onClick={closeModal}>
-                &times;
-              </button>
-              {renderContent()}
-            </div>
-          </div>
-          )} 
-          </div>
-        <div className="right-column">
-        <img src="/sushi-tech-2.jpg" alt="Sushi Tech" />
-        </div>
-    </div>
-    </div>
-  );
-}        
-
-export default Reservas;
