+++ src/pages/reservas/index.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Reservas from '@/components/reservas/reservas';
+
+const Reserva = () => {
+	return (
+		<>
+			<Reservas />
+		</>
+	);
+};
+
+export default Reserva;
