+++ src/styles/global.css
@@ -13,173 +13,3 @@ body {
 	color: #ffffff;
 	background-color: #000000;
 }
-
-/* Estilos para o modal */
-.modal-overlay {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	position: fixed;
-	top: 0;
-	left: 0;
-	right: 0;
-	bottom: 0;
-	background: rgba(0, 0, 0, 0.5);
-  }
-  
-  .modal {
-	color:#000000;
-	background: #fff;
-	border-radius: 10px;
-	max-width: 400px;
-	padding: 20px;
-	text-align: center;
-	position: relative;
-  }
-  
-  .modal-content label {
-	display: block;
-	margin-bottom: 10px;
-  }
-  
-  .modal-content input {
-	width: 100%;
-	padding: 5px;
-	margin-bottom: 10px;
-	border: 1px solid #ccc;
-	border-radius: 5px;
-  }
-  
-  .open-modal-button {
-	background-color: #ff6600;
-	font-family: 'Poppins', sans-serif;
-	color: #fff;
-	border-color: #ff6600;
-	padding: 10px 20px;
-	border-radius: 5px;
-	bottom: 10px;
-	left: 50%;
-  }
-
-  .open-modal-button:hover {
-	color: #fff;
-	font-family: 'Poppins', sans-serif;
-	background-color: #c95000;
-	border-color: #c95000;
-  }
-  .close-modal-button {
-	position: absolute;
-	top: 10px;
-	right: 10px;
-	font-size: 20px;
-	cursor: pointer;
-  }
-  
-  .find-tables-button {
-	background-color: #ff6600;
-	color: #fff;
-	font-family: 'Poppins', sans-serif;
-	border-color: #ff6600;
-	padding: 10px 20px;
-	border-radius: 5px;
-	cursor: pointer;
-	position: center;
-	bottom: 10px;
-	left: 50%;
-  }
-  
-  .find-tables-button:hover {
-	color: #fff;
-	background-color: #c95000;
-	border-color: #c95000;
-  }
-  
-  /* Estilos para a seleção de areas */
-.areas-container {
-	display: grid;
-	grid-template-columns: repeat(auto-fill, minmax(100px, 1fr));
-	grid-gap: 10px;
-	max-width: 400px;
-	margin: 0 auto;
-	text-align: center;
-  }
-  
-  .area {
-	background: #ff6600;
-	color: #fff;
-	border: 2px solid #c95000;
-	padding: 10px;
-	border-radius: 10px;
-	cursor: pointer;
-	font-weight: bold;
-	font-size: 16px;
-  }
-  
-  .area.selected {
-	background: #c95000; /* Cor de destaque para a area selecionada */
-	border: 1px solid #000000;
-  }
-  
-  .confirmar-area-button {
-	background-color: #ff6600;
-	color: #fff;
-	font-family: 'Poppins', sans-serif;
-	border-color: #ff6600;
-	padding: 10px 20px;
-	border-radius: 5px;
-	position: center;
-  }
-  
-  .confirmar-area-button:hover {
-	color: #fff;
-	background-color: #c95000;
-	border-color: #c95000;
-  }
-
-  .confirmar-reserva-button {
-	background-color: #ff6600;
-	color: #fff;
-	font-family: 'Poppins', sans-serif;
-	border-color: #ff6600;
-	padding: 10px 20px;
-	border-radius: 5px;
-	cursor: pointer;
-	position: center;
-	bottom: 10px;
-	left: 50%;
-  }
-  
-  .confirmar-reserva-button:hover {
-	color: #fff;
-	background-color: #c95000;
-	border-color: #c95000;
-  }
-
-  .custom-layout {
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-	padding-bottom: 100px;
-  }
-  
-  .left-column {
-	flex: 1;
-	padding: 100px;
-	text-align: center;
-  }
-  
-  .right-column {
-	flex: 1;
-	padding: 100px;
-	text-align: right;
-  }
-  
-  .right-column img {
-	max-width: 100%; /* A imagem não deve exceder a largura da coluna */
-	height: auto; /* Mantém a proporção da imagem */
-  }
-  
-  .title{
-	text-align: center;
-	padding-top: 50px;
-	}
